Homework 7. Изменение структуры проекта
