+++ app/src/main/java/com/github/pashmentov96/fragments/AppSQLiteOpenHelper.java
@@ -3,7 +3,6 @@ package com.github.pashmentov96.fragments;
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
-import android.provider.BaseColumns;
 
 public class AppSQLiteOpenHelper extends SQLiteOpenHelper {
