+++ app/src/main/java/com/github/pashmentov96/fragments/MainActivity.java
@@ -16,8 +16,9 @@ public class MainActivity extends AppCompatActivity implements ViewHolderListene
         showProfileFragment(id);
     }
 
-    private static String PROFILE_ON_DETAIL_CONTAINER = "Profile on detail container";
-    private static String PROFILE_ON_MAIN_CONTAINER = "Profile on main container";
+    private static final String PROFILE_ON_DETAIL_CONTAINER = "Profile on detail container";
+    private static final String PROFILE_ON_MAIN_CONTAINER = "Profile on main container";
+    private static final String LOG = "MyLogs";
     private static int lastId;
 
     public static Intent getIntent(@NonNull final Context context) {
@@ -51,7 +52,7 @@ public class MainActivity extends AppCompatActivity implements ViewHolderListene
     }
 
     public void showProfileFragment(int id) {
-        Log.d("MyLogs", "ProfileId = " + id);
+        Log.d(LOG, "ProfileId = " + id);
         lastId = id;
         View detailView = findViewById(R.id.detailContainer);
         if (detailView == null) {
